# src/swagger/swagger_output.yaml
...
openapi: 3.0.0
info:
  version: 0.0.0
  title: Vinymatic API
  description: ""
host: localhost:5000
basePath: ""
...
schemes: []
securityDefinitions: {}
consumes: []
produces: []
paths:
...
  "/audios":
    get:
      tags:
      - Audio
      description: ""
      parameters:
      - $ref: "#/components/parameters/pageParam"
      - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audios"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/audio/{id}":
    get:
      tags:
      - Audio
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audio"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
    patch:
      tags:
      - Audio
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      - $ref: "#/components/parameters/typeAudioParams"
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/requestBodies/AddAudio"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audio"
        "304":
          description: Not Modified
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/song/{id}":
    post:
      tags:
      - Audio
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      - $ref: "#/components/parameters/typeAudioParams"
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/requestBodies/AddSongFile"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: 
                    type: integer
                  title: 
                    type: string
                  audioUrl: 
                    type: string
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/song/{typeAudio}/{key}":
    get:
      tags:
      - Audio
      description: "Le chemin correspond au chemin qui mène au ficher dans AWS S3"
      produces:
      - audio/*
      parameters:
      - name: typeAudio
        description: "Correspond au type de l'audio. (valeurs possible: ___songs___)."
        example: songs        
        in: path
        required: true
        schema:
          type: string
      - name: key
        description: "Correspond au nom du fichier audio."
        in: path
        required: true
        schema:
          type: string
      - name: isGetDuration
        description: "Récupérer la ___duration___ du ficher audio stocké dans AWS S3"
        example: true
        in: query
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            audio/*:
              type: string
              format: base64
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
# Format ✔︎
...
  "/image/{imageFrom}/{typeImage}/{key}":
    get:
      tags:
      - Image
      description: "Le chemin correspond au chemin qui mène au ficher dans AWS S3"
      parameters:
      - $ref: "#/components/parameters/imageFrom"
      - $ref: "#/components/parameters/typeImage"
      - $ref: "#/components/parameters/keyFile"
      responses:
        "200":
          description: OK
          content:
            image/*:
              schema: 
                type: string
                format: binary
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
    # delete:
    #   tags:
    #   - Image
    #   description: ""
    #   parameters:
    #   - $ref: "#/components/parameters/imageFrom"
    #   - $ref: "#/components/parameters/typeImage"
    #   - $ref: "#/components/parameters/keyFile"
    #   responses:
    #     "204":
    #       description: Deleted
    #     "401":
    #       description: Unauthorized
    #     "404":
    #       description: Not Found
    #     "500":
    #       description: Internal Server Error
    #   security:
    #   - bearerAuth: []
# Label ✔︎
...
  "/video/{id}":
    get:
      tags:
      - Video
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MinOfVideo"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
    patch:
      tags:
      - Video
      description: ""
      parameters:
        - $ref: "#/components/parameters/idParam"
        - name: idArtist
          in: query
          schema:
            type: integer
        - name: idUser
          in: query
          schema:
            type: integer
      requestBody:
        content:
          multipart/form-data:
            schema:
              allOf:
              - $ref: "#/components/requestBodies/AddVideo"
              - type: object
                properties: 
                  image:
                    type: string
                    format: binary
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MinOfVideo"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
    delete:
      tags:
      - Video
      description: ""
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
# Vinyl ✔︎
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    pageParam:
      name: page
      in: query
      schema:
        type: integer
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
...
    typeAudioParams:
      name: typeAudio
      description: "Correspond au type de l'audio. (valeurs possible: ___song___)."
      in: query
      required: true
      schema:
        type: string
...
    idParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    imageFrom:
      name: imageFrom
      description: "Correspond à l'entiter dans laquelle on veut récupérer l'image. (exemple: ___artists___, si on veux l'image d'un artiste)"
      in: path
      required: true
      schema:
        type: string
    typeImage:
      name: typeImage
      description: "Correspond au type de l'image. (valeurs possible: ___larges___ / ___thumbnails___ | Pour: __imageFrom__ = \"user\" => ___profilImage___ / ___coverImage___)."
      in: path
      required: true
      schema:
        type: string
    keyFile:
      name: key
      description: "Correspond au nom de l'image."
      in: path
      required: true
      schema:
        type: string
  requestBodies:
...
    AddVideo:
      type: object
      properties:
        title:
          type: string
        description:
          type: text
...
    AddAudio:
      type: object
      properties:
        title: 
          type: string
        description: 
          type: text
        image: 
          type: string
          format: binary
        duration: 
          type: integer
        position: 
          type: string
        type: 
          type: string
        # credits: 
        #   type: object
        songFile: 
          type: string
          format: binary
...
    AddSongFile: 
      type: object
      properties:
        songFile:
          type: string
          format: base64
...
  schemas:
...
    MinOfArtist:
      type: object
      properties:
        id: 
          type: integer
        idArtist: 
          type: integer
        name:
          type: string
        thumbnail:
          type: string
        artistUrl:
          type: string
        ressourceUrl:
          type: string
    MinOfAudio:
      type: object
      properties:
        id: 
          type: integer
        title: 
          type: string
        image: 
          type: string
        duration: 
          type: integer
        audioUrl: 
          type: string
        resourceUrl: 
          type: string
...
    MinOfPlaylist:
      type: object
      properties:
        id: 
          type: integer
        title: 
          type: string
        duration: 
          type: integer
        image: 
          type: string
        playlistUrl: 
          type: string
        resourceUrl: 
          type: string
...
    MinOfVideo:
      type: object
      properties:
        id: 
          type: integer
        title:
          type: string  
        description:
          type: text 
        videoUrl:
          type: string  
    MinOfVinyl:
      type: object
      properties:
        id: 
          type: integer
        idRelease: 
          type: integer  
        title:
          type: string  
        thumbnail:
          type: string  
        vinylUrl:
          type: string
        resourceUrl:
          type: string
  
  # Array of
...
    ArrayOfAudio:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfAudio"
...
    ArrayOfPlaylist:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfPlaylist"
...
    Audio:
      type: object
      properties: 
        audio:
          allOf:
          - $ref: "#/components/schemas/MinOfAudio"
          - type: object
            properties:
              description: 
                type: text
              position: 
                type: string
              type: 
                type: string
              credits: 
                type: object
                properties: 
                  creditRole:
                    type: string
                  creditType:
                    type: string
                  creditArtistId:
                    type: integer
                  creditArtistName:
                    type: string
              Vinyl: 
                $ref: "#/components/schemas/MinOfVinyl"
              Artist: 
                $ref: "#/components/schemas/MinOfArtist"
              PlayList: 
                $ref: "#/components/schemas/ArrayOfPlaylist"
...
    Audios:
      type: object
      properties:
        audios:
          allOf:
          - $ref: "#/components/schemas/ArrayOfAudio"
...